+++ spring-cloud-tencent-examples/quickstart-example/quickstart-caller-service/src/main/java/com/tencent/cloud/quickstart/caller/circuitbreaker/CircuitBreakerController.java
@@ -23,7 +23,6 @@ import reactor.core.publisher.Mono;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.cloud.client.circuitbreaker.CircuitBreakerFactory;
-import org.springframework.cloud.client.circuitbreaker.ReactiveCircuitBreakerFactory;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -55,9 +54,6 @@ public class CircuitBreakerController {
 
 	@Autowired
 	private CircuitBreakerFactory circuitBreakerFactory;
-
-	@Autowired
-	private ReactiveCircuitBreakerFactory reactiveCircuitBreakerFactory;
 
 	@Autowired
 	private WebClient.Builder webClientBuilder;
@@ -146,11 +142,6 @@ public class CircuitBreakerController {
 				.get()
 				.uri("/quickstart/callee/circuitBreak")
 				.retrieve()
-				.bodyToMono(String.class)
-				.transform(it ->
-						reactiveCircuitBreakerFactory
-								.create(MetadataContext.LOCAL_NAMESPACE + "QuickstartCalleeService#/quickstart/callee/circuitBreak#http#GET")
-								.run(it, throwable -> Mono.just("fallback: trigger the refuse for service callee"))
-				);
+				.bodyToMono(String.class);
 	}
 }