+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 ---
 
 - [feat: implement circuit breaker in enhance plugin, support listen config group, support refresh single config in refresh_context mode.](https://github.com/Tencent/spring-cloud-tencent/pull/1566)
+- [feat:support polaris event.](https://github.com/Tencent/spring-cloud-tencent/pull/1571)